# wmma: widen TC usage in search by using PADTO on TC axes when possible