Currently works on METAL and CUDA after the latest linearizer fixes, but on HSA there are weird issues with random small odd sized matrix shapes.  The last row or few rows will have random number of elements not be correct and the amount will generally be the same between runs.

To test / repro use:

```sh
PYTHONPATH=. DEBUG=4 DEBUG_VALUES=1 M=16 N=9 K=31 HALF=1 TC_OPT=2 python3 ./extra/gemm/simple_matmul.py
```

Note that while `(16,9,31)` doesn't work, `(16,9,32)` and `(16,17,31)` are fine.  The kernel code looks fine so it seems like memory load or sync issues (however, inserting barriers everywhere didn't make a difference).

For a more complete search use:

```sh
PYTHONPATH=. HSA=1 M_START=12 M_STOP=20 M_STEP=1 N_START=12 N_STOP=20 N_STEP=1 K_START=28 K_STOP=36 K_STEP=1 HALF=1 TC_OPT=2 python3 ./extra/gemm/fuzz_matmul.py
```