@flammit 2024-04-19
For posterity, @chenyuxyz notes that the disabled fuzz test for AMD in the benchmarks is due to an AMD compiler error.