@flammit 2024-04-19
here's the benchmarks you asked for:

baseline:
```
flam@tiny13:~/tinygrad$ time PYTHONPATH=. HSA=1 STEPS=10 DEFAULT_FLOAT=HALF python3 examples/hlb_cifar10.py
shuffling training dataset in 1374.67 ms (epoch=0)
  0 8681.66 ms run, 8679.43 ms python,    2.24 ms HSA, 1198.00 loss, 0.001494 LR, 0.42 GB used,     77.58 GFLOPS,    673.57 GOPS
  1 1787.44 ms run, 1786.63 ms python,    0.81 ms HSA, 1196.00 loss, 0.002989 LR, 2.93 GB used,    375.93 GFLOPS,    671.95 GOPS
  2  216.25 ms run,    2.18 ms python,  214.06 ms HSA, 1893.00 loss, 0.003132 LR, 2.93 GB used,   3107.31 GFLOPS,    671.95 GOPS
  3  208.20 ms run,    1.80 ms python,  206.40 ms HSA, 2690.00 loss, 0.002697 LR, 2.93 GB used,   3227.52 GFLOPS,    671.95 GOPS
  4  209.38 ms run,    1.78 ms python,  207.60 ms HSA, 1787.00 loss, 0.002262 LR, 2.93 GB used,   3209.23 GFLOPS,    671.95 GOPS
  5  209.29 ms run,    1.76 ms python,  207.53 ms HSA, 1641.00 loss, 0.001827 LR, 2.93 GB used,   3210.64 GFLOPS,    671.95 GOPS
  6  209.31 ms run,    1.78 ms python,  207.53 ms HSA, 1413.00 loss, 0.001391 LR, 2.93 GB used,   3210.39 GFLOPS,    671.95 GOPS
  7  209.34 ms run,    1.81 ms python,  207.53 ms HSA, 1318.00 loss, 0.000957 LR, 2.93 GB used,   3209.94 GFLOPS,    671.95 GOPS
  8  209.27 ms run,    2.29 ms python,  206.98 ms HSA, 1313.00 loss, 0.000521 LR, 2.93 GB used,   3210.95 GFLOPS,    671.95 GOPS
  9  207.93 ms run,    1.76 ms python,  206.17 ms HSA, 1267.00 loss, 0.000086 LR, 2.93 GB used,   3231.57 GFLOPS,    671.95 GOPS
shuffling test dataset in 146.45 ms (epoch=0)
eval     1061/10240 10.36%,    2.45 val_loss STEP=10 (in 2080.42 ms)

real	0m19.193s
user	0m18.092s
sys	0m7.397s
```

With TC_OPT=2:
```
flam@tiny13:~/tinygrad$ time PYTHONPATH=. HSA=1 STEPS=10 DEFAULT_FLOAT=HALF TC_OPT=2 python3 examples/hlb_cifar10.py
shuffling training dataset in 1882.42 ms (epoch=0)
  0 19295.01 ms run, 19292.12 ms python,    2.90 ms HSA, 1198.00 loss, 0.001494 LR, 0.42 GB used,     34.89 GFLOPS,    673.24 GOPS
  1 4845.48 ms run, 4844.44 ms python,    1.04 ms HSA, 1196.00 loss, 0.002989 LR, 2.93 GB used,    138.61 GFLOPS,    671.63 GOPS
  2  204.37 ms run,    2.22 ms python,  202.14 ms HSA, 1894.00 loss, 0.003132 LR, 2.93 GB used,   3286.40 GFLOPS,    671.63 GOPS
  3  197.65 ms run,    1.85 ms python,  195.80 ms HSA, 2688.00 loss, 0.002697 LR, 2.93 GB used,   3398.05 GFLOPS,    671.63 GOPS
  4  198.08 ms run,    1.83 ms python,  196.25 ms HSA, 1787.00 loss, 0.002262 LR, 2.93 GB used,   3390.77 GFLOPS,    671.63 GOPS
  5  197.57 ms run,    1.78 ms python,  195.79 ms HSA, 1607.00 loss, 0.001827 LR, 2.93 GB used,   3399.48 GFLOPS,    671.63 GOPS
  6  198.22 ms run,    1.79 ms python,  196.43 ms HSA, 1395.00 loss, 0.001391 LR, 2.93 GB used,   3388.33 GFLOPS,    671.63 GOPS
  7  199.09 ms run,    2.21 ms python,  196.88 ms HSA, 1286.00 loss, 0.000957 LR, 2.93 GB used,   3373.53 GFLOPS,    671.63 GOPS
  8  198.97 ms run,    2.14 ms python,  196.83 ms HSA, 1316.00 loss, 0.000521 LR, 2.93 GB used,   3375.57 GFLOPS,    671.63 GOPS
  9  198.65 ms run,    1.82 ms python,  196.83 ms HSA, 1261.00 loss, 0.000086 LR, 2.93 GB used,   3381.02 GFLOPS,    671.63 GOPS
shuffling test dataset in 224.29 ms (epoch=0)
eval     1223/10240 11.94%,    2.43 val_loss STEP=10 (in 2681.26 ms)

real	0m33.445s
user	0m29.671s
sys	0m8.484s
```


Here's the beam=2:

```
flam@tiny13:~/tinygrad$ time PYTHONPATH=. HSA=1 STEPS=10 DEFAULT_FLOAT=HALF BEAM=2 python3 examples/hlb_cifar10.py
downloading and extracting CIFAR...
shuffling training dataset in 21458.93 ms (epoch=0)
  0 482827.31 ms run, 482826.71 ms python,    0.60 ms HSA, 1198.00 loss, 0.001494 LR, 0.42 GB used,      1.40 GFLOPS,    676.96 GOPS
  1 77799.60 ms run, 77799.14 ms python,    0.46 ms HSA, 1196.00 loss, 0.002989 LR, 3.60 GB used,      8.68 GFLOPS,    675.36 GOPS
  2   61.20 ms run,    2.26 ms python,   58.94 ms HSA, 1894.00 loss, 0.003132 LR, 3.60 GB used,  11035.95 GFLOPS,    675.36 GOPS
  3   54.96 ms run,    1.87 ms python,   53.09 ms HSA, 2686.00 loss, 0.002697 LR, 3.60 GB used,  12288.22 GFLOPS,    675.36 GOPS
  4   54.76 ms run,    1.86 ms python,   52.90 ms HSA, 1785.00 loss, 0.002262 LR, 3.60 GB used,  12332.40 GFLOPS,    675.36 GOPS
  5   55.00 ms run,    1.83 ms python,   53.16 ms HSA, 1618.00 loss, 0.001827 LR, 3.60 GB used,  12280.19 GFLOPS,    675.36 GOPS
  6   54.94 ms run,    1.81 ms python,   53.12 ms HSA, 1394.00 loss, 0.001391 LR, 3.60 GB used,  12293.11 GFLOPS,    675.36 GOPS
  7   54.97 ms run,    1.81 ms python,   53.17 ms HSA, 1310.00 loss, 0.000957 LR, 3.60 GB used,  12285.41 GFLOPS,    675.36 GOPS
  8   54.85 ms run,    1.76 ms python,   53.08 ms HSA, 1307.00 loss, 0.000521 LR, 3.60 GB used,  12313.41 GFLOPS,    675.36 GOPS
  9   54.85 ms run,    1.79 ms python,   53.06 ms HSA, 1259.00 loss, 0.000086 LR, 3.60 GB used,  12312.78 GFLOPS,    675.36 GOPS
shuffling test dataset in 1801.48 ms (epoch=0)
eval     1066/10240 10.41%,    2.44 val_loss STEP=10 (in 12690.45 ms)

real	9m48.498s
user	85m36.636s
sys	74m20.487s
```