@flammit 2024-04-19
> do you have beam=2 for master?

hrm... it is slower even compared to master beam=2.  not clear to me why except for not a great search algorithm.

```
flam@tiny13:~/tinygrad$ time PYTHONPATH=. HSA=1 STEPS=10 DEFAULT_FLOAT=HALF BEAM=2 python3 examples/hlb_cifar10.py
shuffling training dataset in 19376.25 ms (epoch=0)
  0 506001.03 ms run, 506000.44 ms python,    0.59 ms HSA, 1198.00 loss, 0.001494 LR, 0.42 GB used,      1.34 GFLOPS,    676.94 GOPS
  1 76135.38 ms run, 76134.91 ms python,    0.46 ms HSA, 1196.00 loss, 0.002989 LR, 3.60 GB used,      8.87 GFLOPS,    675.32 GOPS
  2   57.25 ms run,    2.33 ms python,   54.92 ms HSA, 1893.00 loss, 0.003132 LR, 3.60 GB used,  11795.22 GFLOPS,    675.32 GOPS
  3   50.84 ms run,    1.91 ms python,   48.93 ms HSA, 2692.00 loss, 0.002697 LR, 3.60 GB used,  13284.22 GFLOPS,    675.32 GOPS
  4   50.88 ms run,    1.95 ms python,   48.93 ms HSA, 1789.00 loss, 0.002262 LR, 3.60 GB used,  13272.48 GFLOPS,    675.32 GOPS
  5   51.00 ms run,    1.84 ms python,   49.16 ms HSA, 1640.00 loss, 0.001827 LR, 3.60 GB used,  13240.49 GFLOPS,    675.32 GOPS
  6   50.96 ms run,    1.83 ms python,   49.13 ms HSA, 1411.00 loss, 0.001391 LR, 3.60 GB used,  13252.87 GFLOPS,    675.32 GOPS
  7   50.92 ms run,    1.85 ms python,   49.07 ms HSA, 1317.00 loss, 0.000957 LR, 3.60 GB used,  13261.89 GFLOPS,    675.32 GOPS
  8   51.04 ms run,    1.83 ms python,   49.21 ms HSA, 1310.00 loss, 0.000521 LR, 3.60 GB used,  13231.15 GFLOPS,    675.32 GOPS
  9   51.07 ms run,    1.84 ms python,   49.23 ms HSA, 1262.00 loss, 0.000086 LR, 3.60 GB used,  13222.12 GFLOPS,    675.32 GOPS
shuffling test dataset in 3300.60 ms (epoch=0)
eval     1037/10240 10.13%,    2.44 val_loss STEP=10 (in 16446.25 ms)

real	10m7.960s
user	87m43.157s
sys	72m47.920s
```