@chenyuxyz 2024-04-19
i think HSA compiler is weird and the result can be incorrect, benchmark with cuda?